+++ Source/App/App.cs
@@ -220,7 +220,7 @@ namespace WindowsVirtualDesktopHelper {
 				this.UIUpdateIconForVDDisplayNumber(this.CurrentSystemThemeName, this.CurrentVDDisplayNumber, this.CurrentVDDisplayName);
 				this.UIUpdateIconForVDDisplayName(this.CurrentSystemThemeName, this.CurrentVDDisplayName);
 				this.UIUpdateNextPrevIconVisibility(this.CurrentSystemThemeName);
-				// Show overlay
+				// Show notification overlay
 				if(Settings.GetBool("feature.showDesktopSwitchOverlay")) {
 					this.AppForm.Invoke((Action)(() => {
 						SwitchNotificationForm.CloseAllNotifications(this.AppForm);
@@ -239,6 +239,8 @@ namespace WindowsVirtualDesktopHelper {
 						}
 					}));
 				}
+				// Update permanent overlay
+				UpdateStatusOverlayWindows();
 				// Restore focus
 				if(Settings.GetBool("feature.restorePreviousWindowFocus")) {
 					try {
@@ -300,6 +302,27 @@ namespace WindowsVirtualDesktopHelper {
 					this.SwitchToDesktop(_desktopNumberHistory[i]);
 					return;
 				}
+			}
+		}
+
+		public void UpdateStatusOverlayWindows() {
+			if(Settings.GetBool("feature.showDesktopStatusOverlay")) {
+				this.AppForm.Invoke((Action)(() => {
+					OverlayForm.CloseAllNotifications(this.AppForm);
+					if(Settings.GetBool("feature.showDesktopStatusOverlay.showOnAllMonitors")) {
+						for(var i = 0; i < Screen.AllScreens.Length; i++) {
+							var form = new OverlayForm(i);
+							form.LabelText = this.CurrentVDDisplayName;
+							form.Show();
+						}
+					} else {
+						var form = new OverlayForm();
+						form.LabelText = this.CurrentVDDisplayName;
+						form.Show();
+					}
+				}));
+			} else {
+				OverlayForm.CloseAllNotifications(this.AppForm);
 			}
 		}
 
@@ -683,7 +706,7 @@ namespace WindowsVirtualDesktopHelper {
 
 		public void UIUpdateIconForVDDisplayNumber(string theme, uint number, string name) {
 			number++;
-			this.AppForm.notifyIconNumber.Icon = Util.Icons.GenerateNotificationIcon(number.ToString(), theme, this.AppForm.DeviceDpi, false, FontStyle.Bold);
+			this.AppForm.notifyIconNumber.Icon = Util.Icons.GenerateNotificationIcon(number.ToString(), theme, this.AppForm.DeviceDpi, false);
 		}
 
 		public void UIUpdateIconForVDDisplayName(string theme, string name) {
@@ -701,8 +724,8 @@ namespace WindowsVirtualDesktopHelper {
 				var hasNextDesktop = count != 0 && App.Instance.CurrentVDDisplayNumber != count;
 				var hasPrevDesktop = App.Instance.CurrentVDDisplayNumber != 0;
 				// Update prev/next icons
-				this.AppForm.notifyIconPrev.Icon = Util.Icons.GenerateNotificationIcon(prevChar, theme, this.AppForm.DeviceDpi, true, FontStyle.Regular, hasPrevDesktop ? 1.0f : Settings.GetDouble("theme.icons.disabledOpacity"));
-				this.AppForm.notifyIconNext.Icon = Util.Icons.GenerateNotificationIcon(nextChar, theme, this.AppForm.DeviceDpi, true, FontStyle.Regular, hasNextDesktop ? 1.0f : Settings.GetDouble("theme.icons.disabledOpacity"));
+				this.AppForm.notifyIconPrev.Icon = Util.Icons.GenerateNotificationIcon(prevChar, theme, this.AppForm.DeviceDpi, true, hasPrevDesktop ? 1.0f : Settings.GetDouble("theme.icons.disabledOpacity"));
+				this.AppForm.notifyIconNext.Icon = Util.Icons.GenerateNotificationIcon(nextChar, theme, this.AppForm.DeviceDpi, true, hasNextDesktop ? 1.0f : Settings.GetDouble("theme.icons.disabledOpacity"));
 				// Show or hide?
 				if(Settings.GetBool("feature.showPrevNextIcons.automaticallyHidePrevNextOnBounds")) {
 					this.AppForm.notifyIconNext.Visible = hasNextDesktop;