+++ Source/App/Settings.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.Globalization;
 using System.Linq;
 using System.Runtime.InteropServices;
@@ -34,21 +35,30 @@ namespace WindowsVirtualDesktopHelper {
 
 			// Theme
 			RegisterDefault("theme.icons.disabledOpacity", 0.5, "Defines the opacity to use for icons which are disabled.");
-			RegisterDefault("theme.icons.font", "Segoe UI", "Defines the font name to use for the icons (for regular numbers, characters).");
+			RegisterDefault("theme.icons.font", "Segoe UI", "Defines the font name to use for the icons (for regular numbers, characters). If a specific style is to be used, then one can append 'Bold', 'Italic', 'Regular' after a comma and the font name - for example 'Arial, Bold'.");
 			RegisterDefault("theme.icons.emojiFont", "Segoe UI Symbol", "Defines the font name to use for emoji icons.");
 			RegisterDefault("theme.icons.symbolsFont", "Segoe UI Symbol", "Defines the font name to use for symbol icons.");
 			RegisterDefault("theme.icons.iconBG.dark", "black");
 			RegisterDefault("theme.icons.iconFG.dark", "white");
 			RegisterDefault("theme.icons.iconBG.light", "white");
 			RegisterDefault("theme.icons.iconFG.light", "black");
-			RegisterDefault("theme.overlay.width", 900, "With width in pixels of the overlay.");
-			RegisterDefault("theme.overlay.height", 430, "With height in pixels of the overlay.");
-			RegisterDefault("theme.overlay.font", "Segoe UI Light", "Defines the font name to use for the overlay.");
-			RegisterDefault("theme.overlay.fontSize", 30, "Defines the font size to use for the overlay.");
+			RegisterDefault("theme.overlay.width", 900, "With width in pixels of the switch overlay.");
+			RegisterDefault("theme.overlay.height", 430, "With height in pixels of the switch overlay.");
+			RegisterDefault("theme.overlay.font", "Segoe UI Light", "Defines the font name to use for the switch overlay.");
+			RegisterDefault("theme.overlay.fontSize", 30, "Defines the font size to use for the switch overlay.");
 			RegisterDefault("theme.overlay.overlayBG.dark", "black");
 			RegisterDefault("theme.overlay.overlayFG.dark", "white");
 			RegisterDefault("theme.overlay.overlayBG.light", "black");
 			RegisterDefault("theme.overlay.overlayFG.light", "white");
+			RegisterDefault("theme.status.width", 250, "With width in pixels of the status overlay.");
+			RegisterDefault("theme.status.height", 40, "With height in pixels of the status overlay.");
+			RegisterDefault("theme.status.offset", 0, "With height in pixels of the status overlay.");
+			RegisterDefault("theme.status.font", "Segoe UI Light", "Defines the font name to use for the status overlay.");
+			RegisterDefault("theme.status.fontSize", 12, "Defines the font size to use for the status overlay.");
+			RegisterDefault("theme.status.overlayBG.dark", "black");
+			RegisterDefault("theme.status.overlayFG.dark", "white");
+			RegisterDefault("theme.status.overlayBG.light", "black");
+			RegisterDefault("theme.status.overlayFG.light", "white");
 
 			// Feature: splash
 			RegisterDefault("feature.showSplashScreen", true, "If enabled, a splash screen is shown on startup of the app. Overlays must be enabled.");
@@ -68,6 +78,13 @@ namespace WindowsVirtualDesktopHelper {
 			RegisterDefault("feature.showDesktopSwitchOverlay.translucent", true);
 			RegisterDefault("feature.showDesktopSwitchOverlay.showOnAllMonitors", true);
 			RegisterDefault("feature.showDesktopSwitchOverlay.position", "middlecenter");
+
+			// Feature: showDesktopStatusOverlay
+			RegisterDefault("feature.showDesktopStatusOverlay", false);
+			RegisterDefault("feature.showDesktopStatusOverlay.animate", true);
+			RegisterDefault("feature.showDesktopStatusOverlay.translucent", true);
+			RegisterDefault("feature.showDesktopStatusOverlay.showOnAllMonitors", true);
+			RegisterDefault("feature.showDesktopStatusOverlay.position", "topcenter");
 
 			// Feature: useHotKeyToJumpToDesktopNumber
 			RegisterDefault("feature.useHotKeyToJumpToDesktopNumber", false);
@@ -195,6 +212,33 @@ namespace WindowsVirtualDesktopHelper {
 			var ret = _get(key, defaultValue);
 			if(ret == null) return null;
 			return ret as string;
+		}
+
+		public static string GetFontName(string key, string defaultValue = null) {
+			var val = GetString(key, defaultValue); // can include style, like "Segoe UI, Bold"
+			if(val == null) return defaultValue;
+			if(val.Contains(",")) {
+				var segs = val.Split(',');
+				return segs.First().Trim(); // Return the first segment as the font name
+			} else {
+				return val.Trim();
+			}
+		}
+
+		public static FontStyle GetFontStyle(string key, FontStyle defaultValue = FontStyle.Regular) {
+			var val = GetString(key); // can include style, like "Segoe UI, Bold"
+			if(val == null) return defaultValue;
+			if(val.Contains(",")) {
+				var segs = val.Split(',');
+				var styleStr = segs.Last().Trim().ToLower();
+				if(styleStr == "bold") return FontStyle.Bold;
+				else if(styleStr == "italic") return FontStyle.Italic;
+				else if(styleStr == "underline") return FontStyle.Underline;
+				else if(styleStr == "strikeout") return FontStyle.Strikeout;
+				else return FontStyle.Regular; // Default case
+			} else {
+				return defaultValue;
+			}
 		}
 
 		public static void SetString(string key, string value) {